CBSP-602: Refactor to extract settings to separate file
Editable settings are now contained within `settings.rb`. This is loaded near the start of `Top_Level_Vagrantfile`, but after `OPERATING_SYSTEM` and `VERSION` are evaluated as `VERSION` is required by some of the URLs. All variables have been changed to global constants as they are not expected to be changed at runtime.